[refactor] 페이지 재구성
 - content용 디렉토리 추가 (src/views/contents)
 - header용 디렉토리 추가 (src/views/header)
 - footer용 디렉토리 추가 (src/views/footer)
 - nav용 디렉토리 추가 (src/views/navigation)

 상세:
 - Button이나 List Item을 Component으로 구현해야 할지도 몰라 선제적으로 프로젝트 구조 세분화.
