+++ src/main.js
@@ -1,55 +1,25 @@
 import { Router } from './router.js';
 import { Error } from '/src/component/error.js';
 import { useUserStore } from '/src/store/useUserStore.js';
-import { homePage } from '/src/views/homePage.js';
-import { loginPage } from '/src/views/loginPage.js';
-import { notFoundPage } from '/src/views/notFoundPage.js';
-import { profilePage } from '/src/views/profilePage.js';
+import { homePage } from '/src/views/contents/homePage.js';
+import { loginPage } from '/src/views/contents/loginPage.js';
+import { notFoundPage } from '/src/views/contents/notFoundPage.js';
+import { profilePage } from '/src/views/contents/profilePage.js';
+import { footer } from '/src/views/footer/footer.js';
+import { header } from '/src/views/header/header.js';
+import { nav } from '/src/views/navigation/nav.js';
 
 const router = new Router(document.querySelector('#root'));
 
 // 레이아웃 함수
 const layout = (content) => {
-  const currentPath = window.location.pathname;
-
-  const getNavLinkClass = (path) => {
-    return currentPath === path
-      ? 'text-blue-600 font-bold'
-      : 'text-gray-600 hover:text-blue-600';
-  };
-
   return `
   <div class="bg-gray-100 min-h-screen flex justify-center">
     <div class="max-w-md w-full">
-      <header class="bg-blue-600 text-white p-4 sticky top-0">
-        <h1 class="text-2xl font-bold">항해플러스</h1>
-      </header>
-
-      <nav class="bg-white shadow-md p-2 sticky top-14">
-        <ul class="flex justify-around">
-          <li><a href="/" class="${getNavLinkClass('/')}">홈</a></li>
-          ${
-            useUserStore.isLoggedIn()
-              ? `<li><a href="/profile" class="${getNavLinkClass(
-                  '/profile'
-                )}">프로필</a></li>`
-              : ``
-          }
-          ${
-            useUserStore.isLoggedIn()
-              ? `<li><button id="logout" class="text-gray-600 hover:text-blue-600">로그아웃</button></li>`
-              : `<li><a href="/login" class="${getNavLinkClass(
-                  '/login'
-                )}">로그인</a></li>`
-          }
-        </ul>
-      </nav>
-
+      ${header()}
+      ${nav()}
       ${content}
-
-      <footer class="bg-gray-200 p-4 text-center">
-        <p>&copy; 2024 항해플러스. All rights reserved.</p>
-      </footer>
+      ${footer()}
     </div>
   </div>
 `;
